Split FastAPI routes from main
